requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
appdirs==1.4.4
    # via fs
arrow==1.2.3
    # via jinja2-time
asgiref==3.6.0
    # via
    #   -r requirements/base.txt
    #   django
attrs==22.2.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
    #   pytest
binaryornot==0.4.4
    # via cookiecutter
boto==2.49.0
    # via xblock-sdk
boto3==1.26.40
    # via fs-s3fs
botocore==1.29.40
    # via
    #   boto3
    #   s3transfer
certifi==2022.12.7
    # via requests
chardet==5.1.0
    # via binaryornot
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   code-annotations
    #   cookiecutter
code-annotations==1.3.0
    # via -r requirements/test.in
cookiecutter==2.1.1
    # via xblock-sdk
coverage[toml]==7.0.1
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-model-utils
    #   django-pyfs
    #   djangorestframework
    #   openedx-events
    #   openedx-filters
    #   xblock-sdk
django-model-utils==4.3.1
    # via -r requirements/base.txt
django-pyfs==3.2.0
    # via xblock-sdk
djangorestframework==3.14.0
    # via -r requirements/test.in
edx-opaque-keys[django]==2.3.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
exceptiongroup==1.1.0
    # via pytest
fastavro==1.7.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
fs==2.4.16
    # via
    #   django-pyfs
    #   fs-s3fs
    #   xblock
fs-s3fs==1.1.1
    # via
    #   django-pyfs
    #   xblock-sdk
idna==3.4
    # via requests
iniconfig==1.1.1
    # via pytest
jinja2==3.1.2
    # via
    #   code-annotations
    #   cookiecutter
    #   jinja2-time
jinja2-time==0.2.0
    # via cookiecutter
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lazy==1.5
    # via xblock-sdk
lxml==4.9.2
    # via
    #   xblock
    #   xblock-sdk
markupsafe==2.1.1
    # via
    #   jinja2
    #   xblock
mock==5.0.0
    # via -r requirements/test.in
openedx_events @ git+https://github.com/open-craft/openedx-events.git@77ad965c4ac5157b861a19b8cb6b9240883008ba
    # via -r requirements/base.txt
openedx-filters @ git+https://github.com/open-craft/openedx-filters.git@d0206cf84f2e5e22b1a99d06566d88839d40b9fe
    # via -r requirements/base.txt
packaging==22.0
    # via pytest
pbr==5.11.0
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.0.0
    # via pytest
pymongo==3.13.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pypng==0.20220715.0
    # via xblock-sdk
pytest==7.2.0
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
    #   xblock
python-slugify==7.0.0
    # via
    #   code-annotations
    #   cookiecutter
pytz==2022.7
    # via
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
    #   xblock
pyyaml==6.0
    # via
    #   code-annotations
    #   cookiecutter
    #   xblock
requests==2.28.1
    # via
    #   cookiecutter
    #   xblock-sdk
s3transfer==0.6.0
    # via boto3
simplejson==3.18.0
    # via xblock-sdk
six==1.16.0
    # via
    #   fs
    #   fs-s3fs
    #   python-dateutil
sqlparse==0.4.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==4.1.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
urllib3==1.26.13
    # via
    #   botocore
    #   requests
web-fragments==2.0.0
    # via
    #   xblock
    #   xblock-sdk
webob==1.8.7
    # via
    #   xblock
    #   xblock-sdk
xblock==1.6.1
    # via xblock-sdk
xblock-sdk==0.5.1
    # via -r requirements/test.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
